+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Each change should be sorted in to one of the following categories:
 
 ## [Unreleased]
 ### Added
+- createLobby sdk method
 - updateAchievementProgress sdk method
 - submitCompletedSession sdk method
 - registerGame sdk method